+++ CHANGELOG.md
@@ -7,6 +7,10 @@ BREAKING CHANGES:
 - [DeliverTx] Add new functions (`RevokeAccessorMethod`).
 - [DeliverTx] Check request is closed with valid ial, proof and signature (`AddAccessorMethod`).
 
+IMPROVEMENTS:
+
+- [Query] Add new function (`GetAccessorOwner`).
+
 ## 0.10.1 (October 9, 2018)
 
 IMPROVEMENTS: